Add two phase example and format output.
